spez-backend/requirements.txt
alembic==1.10.4
anyio==4.6.0
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.95.2
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.0
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==1.0.1
sniffio==1.3.1
SQLAlchemy==1.4.46
starlette==0.27.0
typing_extensions==4.12.2
uvicorn==0.22.0
